requirements.txt
streamlit==1.50.0
numpy==1.25.2
Pillow==9.6.0
opencv-python==4.12.0
